Fixes the NT cruciform's odd interaction with stat holders (#5645)
* Starts fixing the NT cruciform cloning making statless humans

Issue: They now make humans with no perks or stat values.

* Sanitizes a bunch of perks from checking holders when there is no holder

Fixes gibbing not calling implant uninstall

Cruciform internal stats update on uninstall

* Removes debug

* Terrible code that should be looked at

+++ code/datums/perks/fate.dm
@@ -5,10 +5,12 @@
 
 /datum/perk/fate/paper_worm/assign(mob/living/carbon/human/H)
 	..()
-	holder.sanity.positive_prob += 20
+	if(holder)
+		holder.sanity.positive_prob += 20
 
 /datum/perk/fate/paper_worm/remove()
-	holder.sanity.positive_prob -= 20
+	if(holder)
+		holder.sanity.positive_prob -= 20
 	..()
 
 /datum/perk/fate/freelancer
@@ -38,13 +40,15 @@
 
 /datum/perk/fate/nihilist/assign(mob/living/carbon/human/H)
 	..()
-	holder.sanity.positive_prob += 10
-	holder.sanity.negative_prob += 20
+	if(holder)
+		holder.sanity.positive_prob += 10
+		holder.sanity.negative_prob += 20
 
 /datum/perk/fate/nihilist/remove()
-	holder.sanity.positive_prob -= 10
-	holder.sanity.negative_prob -= 20
-	holder.stats.removeTempStat(STAT_COG, "Fate Nihilist")
+	if(holder)
+		holder.sanity.positive_prob -= 10
+		holder.sanity.negative_prob -= 20
+		holder.stats.removeTempStat(STAT_COG, "Fate Nihilist")
 	..()
 
 /datum/perk/fate/moralist
@@ -93,14 +97,16 @@
 
 /datum/perk/fate/alcoholic_active/assign(mob/living/carbon/human/H)
 	..()
-	holder.stats.addTempStat(STAT_ROB, 10, INFINITY, "Fate Alcoholic")
-	holder.stats.addTempStat(STAT_TGH, 10, INFINITY, "Fate Alcoholic")
-	holder.stats.addTempStat(STAT_VIG, 10, INFINITY, "Fate Alcoholic")
+	if(holder)
+		holder.stats.addTempStat(STAT_ROB, 10, INFINITY, "Fate Alcoholic")
+		holder.stats.addTempStat(STAT_TGH, 10, INFINITY, "Fate Alcoholic")
+		holder.stats.addTempStat(STAT_VIG, 10, INFINITY, "Fate Alcoholic")
 
 /datum/perk/fate/alcoholic_active/remove()
-	holder.stats.removeTempStat(STAT_ROB, "Fate Alcoholic")
-	holder.stats.removeTempStat(STAT_TGH, "Fate Alcoholic")
-	holder.stats.removeTempStat(STAT_VIG, "Fate Alcoholic")
+	if(holder)
+		holder.stats.removeTempStat(STAT_ROB, "Fate Alcoholic")
+		holder.stats.removeTempStat(STAT_TGH, "Fate Alcoholic")
+		holder.stats.removeTempStat(STAT_VIG, "Fate Alcoholic")
 	..()
 
 /datum/perk/fate/noble
@@ -110,6 +116,9 @@
 
 /datum/perk/fate/noble/assign(mob/living/carbon/human/H)
 	..()
+	if(!holder)
+		return
+	holder.sanity.environment_cap_coeff -= 1
 	if(!holder.family_name)		//Eclipse edit: Family name, not surname
 		qdel(src)
 		return
@@ -139,7 +148,8 @@
 		holder.equip_to_storage_or_drop(W)
 
 /datum/perk/fate/noble/remove()
-	holder.sanity.environment_cap_coeff += 1
+	if(holder)
+		holder.sanity.environment_cap_coeff += 1
 	..()
 
 /datum/perk/fate/rat
@@ -149,10 +159,12 @@
 
 /datum/perk/fate/rat/assign(mob/living/carbon/human/H)
 	..()
-	holder.sanity.max_level -= 10
+	if(holder)
+		holder.sanity.max_level -= 10
 
 /datum/perk/fate/rat/remove()
-	holder.sanity.max_level += 10
+	if(holder)
+		holder.sanity.max_level += 10
 	..()
 
 /datum/perk/fate/rejected_genius
@@ -162,16 +174,18 @@
 
 /datum/perk/fate/rejected_genius/assign(mob/living/carbon/human/H)
 	..()
-	holder.sanity.environment_cap_coeff -= 1
-	holder.sanity.positive_prob_multiplier -= 1
-	holder.sanity.insight_passive_gain_multiplier *= 1.5
-	holder.sanity.max_level -= 20
+	if(holder)
+		holder.sanity.environment_cap_coeff -= 1
+		holder.sanity.positive_prob_multiplier -= 1
+		holder.sanity.insight_passive_gain_multiplier *= 1.5
+		holder.sanity.max_level -= 20
 
 /datum/perk/fate/rejected_genius/remove()
-	holder.sanity.environment_cap_coeff += 1
-	holder.sanity.positive_prob_multiplier += 1
-	holder.sanity.insight_passive_gain_multiplier /= 1.5
-	holder.sanity.max_level += 20
+	if(holder)
+		holder.sanity.environment_cap_coeff += 1
+		holder.sanity.positive_prob_multiplier += 1
+		holder.sanity.insight_passive_gain_multiplier /= 1.5
+		holder.sanity.max_level += 20
 	..()
 
 /datum/perk/fate/oborin_syndrome
@@ -181,12 +195,14 @@
 
 /datum/perk/fate/oborin_syndrome/assign(mob/living/carbon/human/H)
 	..()
-	holder.sanity.max_level += 20
-	spawn(1)
-		holder.update_client_colour() //Handle the activation of the colourblindness on the mob.
+	if(holder)
+		holder.sanity.max_level += 20
+		spawn(1)
+			holder.update_client_colour() //Handle the activation of the colourblindness on the mob.
 
 /datum/perk/fate/oborin_syndrome/remove()
-	holder.sanity.max_level -= 20
+	if(holder)
+		holder.sanity.max_level -= 20
 	..()
 
 /datum/perk/fate/lowborn
@@ -196,8 +212,10 @@
 
 /datum/perk/fate/lowborn/assign(mob/living/carbon/human/H)
 	..()
-	holder.sanity.max_level += 10
+	if(holder)
+		holder.sanity.max_level += 10
 
 /datum/perk/fate/lowborn/remove()
-	holder.sanity.max_level -= 10
+	if(holder)
+		holder.sanity.max_level -= 10
 	..()
